feat: add invoice management features with PDF generation and VAT handling
- Introduced new types and interfaces for invoices, line items, and VAT breakdown in the frontend API.
- Implemented InvoicePDFService to generate PDF invoices from HTML templates in the backend.
- Created comprehensive tests for invoice creation, issuance, and inquiry linkage.
- Added database migration to refactor invoices for draft status and mixed VAT rates.
- Developed frontend API functions for listing, creating, updating, issuing, and downloading invoices.
- Implemented InvoiceStatusBadge component for displaying invoice statuses.
- Created custom hooks for managing invoice-related operations in the frontend.
- Added unit tests for the InvoicesPage component to ensure correct functionality.

+++ backend/app/models/invoice_line_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import Column, ForeignKey, Index, Numeric, String, Text
 from sqlalchemy.orm import relationship
 import uuid
 from app.models.base import Base, TimestampMixin
@@ -14,10 +14,10 @@ class InvoiceLineItem(TimestampMixin, Base):
     invoice_id = Column(String(36), ForeignKey("invoices.id"), nullable=False)
 
     description = Column(Text, nullable=False)
-    quantity = Column(Float, default=1.0, nullable=False)
-    unit_price = Column(Float, nullable=False)
-    vat_rate = Column(Float, default=0.0, nullable=False)
-    line_total_net = Column(Float, nullable=False)
-    line_total_vat = Column(Float, nullable=False)
+    quantity = Column(Numeric(12, 2), default=1.0, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
+    vat_rate = Column(Numeric(5, 2), default=0.0, nullable=False)
+    line_total_net = Column(Numeric(12, 2), nullable=False)
+    line_total_vat = Column(Numeric(12, 2), nullable=False)
 
     invoice = relationship("Invoice", back_populates="line_items")
